Move index.js functionality into logically grouped files
The index.js file was starting to get unruly. This change splits its functionality into four files and imports those back into index.js
